requirements.txt
streamlit
plotly
pandas
python-dotenv
groq
numpy 
tabulate
matplotlib 